backend/requirements.txt
fastapi
uvicorn
pandas
numpy
diffprivlib
python-multipart
firebase-admin
google-cloud-firestore